// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-contract-model.ts
export class EmployeeContractModel {
    employmentContractId: string;
    employeeId: string;
    startDate: Date;
    endDate: Date;
    contractDetails: string;
    contractTypeId: string;
    contractedHours: number;
    hourlyRate: number;
    holidayOrThisYear: string;
    holidayOrFullEntitlement: string;
    isArchived: boolean;
    timeStamp: any;
    currencyId: string;
    currencyName: string;
    symbol: string;

    firstName: string;
    surName: string;
    userName: string;
    email: string;
    contractTypeName: string;
    totalCount: number;
}